- Todas operações de arquivos devem medir todas as metricas - Consigo pegar todos processos em back mas para matá-los tenho que ser root - GPS funcionando, faltava chamar o método que finalizava a thread.

+++ src/edu/performance/test/floatoperation/FloatOperation.java
@@ -18,7 +18,7 @@ import edu.performance.test.PerformanceTestInterface;
 public class FloatOperation extends PerformanceTest<Double> implements PerformanceTestInterface{
 
 	public FloatOperation(PerformanceTestActivity activity) {
-		super(new Double(10000), activity);
+		super(new Double(999983), activity);
 		if(activity != null)
 		activity.executeTest();
 	}
